Began the construction of the AI Player
An AI player needs to know the following: its intelligence level (which
is the game’s difficulty level), and the game it plays. I can represent
those as private attributes and pass the intelligence level to the
constructor and create a little public setter function that attaches the
AI player to the game it plays.

I have constructed three different difficulty levels for the AI: blind,
novice, and master

I've also implemented the MinimaxValue Recursive Function for AI...

(more to come...incomplete)