Confuring the AIs actions
We’ll need the AI action to hold two pieces of information: the position
on the board that it’ll make its move on and the minimax value of the
state that this action will lead to.

Minimax Value pertains to decision theory when it comes to games. Making
sure the AI picks the best possible decision.