More AI (minimax function)
AI uses the minimax value to choose the best action from a list of
available actions, so it’s very reasonable to think that we’d need some
way to sort the actions based on their minimax values.

For that, we provide two public static functions that we can use as a
compare function to pass to the javascript’s Array.prototype.sort
function to sort the list of actions in both ascending and descending
manners.