Developing actual structure of the game
We need to keep into account three pieces of info:

- AI player who plays the game with the human
- The current state of the game
- The status of the game (whether it’s running or ended)

We’ll keep those as public attributes for easy access. We’d also need a
way to move the game from one state to another as the playing continues,
we can implement that with a public function that advances the game to a
given state and checks if the game ends at this state and what is the
result, if the game doesn’t end at this state, the function notifies the
player who plays next to continue. Finally, we’d need a public function
to start the game.