Constructing the different difficulties for the AI (Master and Novice)
Using the minimax algorithm, taking an optimal decision is a trivial
process:

- Enumerate all the possible action that could be taking.
- Calculate the minimax value of each action (that is the minimax value
of the state that each action leads to).
- If X is the one who is taking the decision: choose the action with the
maximum minimax value.

OR

- If O is the one who is taking the decision: choose the action with the
minimum minimax value.

THEN

- Carry out the chosen action.

For the Novice AI
- A Novice player is a player who sometimes takes the optimal move and
some other time he takes a sub-optimal move. We could model that with a
probability, like saying that it takes the optimal move 40% of the time
and the sub-optimal move 60% of the time.

If I want to execute a statement P percent of the time, just generate a
random number between 0 and 100. You only execute that statement if the
generated random number is less than or equal to P.